+++ CHANGELOG.md
@@ -2,14 +2,15 @@
 
 ### master
 
+- refactor(updates): improve readability and memory handling
 - feat(memory): add `SwapCached` info
-- utils(waybar): replace deprecated `trimRight` function with `trimEnd`
-- utils(waybar): replace deprecated `readAll` function with new one
+- utils(waybar): replace deprecated `trimRight()` function with `trimEnd()`
+- utils(waybar): replace deprecated `readAll()` function with new one
 - refactor(ping): remove `TARGET_UPDATE_MS` from config and set to `60`s
 - docs: update readme
 - docs: update changelog
 - ping: remove domain ip update info
-- memory: replace deprecated `readAll` function with new one
+- memory: replace deprecated `readAll()` function with new one
 - updates: update imports for clarity
 
 ## v1.x